%% File Name: pbv-package.Rd
%% File Version: 0.03

\name{pbv-package}
\alias{pbv-package}
\alias{pbv}
\docType{package}
\title{
\packageTitle{pbv}
}


\description{
\packageDescription{pbv}
}

%\details{
%XXXX
%}

\author{
\packageAuthor{pbv}

Maintainer: \packageMaintainer{pbv}
}


\references{
Drezner, Z., & Wesolowsky, G. O. (1990). On the computation of the bivariate normal 
integral. \emph{Journal of Statistical Computation and Simulation, 35}(1-2), 101-107.
\doi{10.1080/00949659008811236}
}


\keyword{package}


%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}

%\examples{
%~~ simple examples of the most important functions ~~
%}
